15: convert undertaker to es6 syntax
