+++ index.js
@@ -1,26 +1,25 @@
 'use strict';
 
-var util = require('util');
 var Undertaker = require('./src/undertaker');
 var vfs = require('vinyl-fs');
 var watch = require('glob-watcher');
 
-function Gulp() {
-  Undertaker.call(this);
+class Gulp extends Undertaker {
+  constructor() {
+    super();
 
-  // Bind the functions for destructuring
-  this.watch = this.watch.bind(this);
-  this.task = this.task.bind(this);
-  this.series = this.series.bind(this);
-  this.parallel = this.parallel.bind(this);
-  this.registry = this.registry.bind(this);
-  this.tree = this.tree.bind(this);
-  this.lastRun = this.lastRun.bind(this);
-  this.src = this.src.bind(this);
-  this.dest = this.dest.bind(this);
-  this.symlink = this.symlink.bind(this);
+    this.watch = this.watch.bind(this);
+    this.task = this.task.bind(this);
+    this.series = this.series.bind(this);
+    this.parallel = this.parallel.bind(this);
+    this.registry = this.registry.bind(this);
+    this.tree = this.tree.bind(this);
+    this.lastRun = this.lastRun.bind(this);
+    this.src = this.src.bind(this);
+    this.dest = this.dest.bind(this);
+    this.symlink = this.symlink.bind(this);
+  }
 }
-util.inherits(Gulp, Undertaker);
 
 Gulp.prototype.src = vfs.src;
 Gulp.prototype.dest = vfs.dest;
