--- src/undertaker/lib/get-task.js
@@ -1,7 +1,0 @@
-'use strict';
-
-function get(name) {
-  return this._registry.get(name);
-}
-
-module.exports = get;
