+++ src/undertaker/validateRegistry.js
